requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
pytest
pytest-mock
pytest-asyncio
pydantic-settings